+++ src/main/java/de/fraunhofer/iosb/ilt/frostclient/model/ModelRegistry.java
@@ -181,4 +181,8 @@ public class ModelRegistry {
         initialised = true;
     }
 
+    public boolean isInitialised() {
+        return initialised;
+    }
+
 }